+++ projects/openttd_vs80.vcproj
@@ -2375,6 +2375,10 @@
 				>
 			</File>
 			<File
+				RelativePath=".\..\src\widgets\framerate_widget.h"
+				>
+			</File>
+			<File
 				RelativePath=".\..\src\widgets\genworld_widget.h"
 				>
 			</File>
